tests(node): refactor node endpoint tests for v26
