+++ src/api/models_new/mod.rs
@@ -1,0 +1,2 @@
+pub mod fee;
+pub mod node;
