+++ tests/fixtures/node-configuration.json
@@ -1,91 +1,77 @@
 {
-  "data": {
-    "nethash": "2a44f340d76ffc3df204c5f38cd355b7496c9065a1ade2ef92071436bd72e867",
-    "token": "DARK",
-    "symbol": "DѦ",
-    "explorer": "https://dexplorer.ark.io",
-    "version": 30,
-    "ports": {
-      "@arkecosystem/core-p2p": 4002,
-      "@arkecosystem/core-api": 4003,
-      "@arkecosystem/core-graphql": 4005
-    },
-    "constants": {
-      "height": 21600,
-      "reward": 200000000,
-      "activeDelegates": 51,
-      "blocktime": 8,
-      "block": {
-        "version": 0,
-        "maxTransactions": 150,
-        "maxPayload": 6300000
-      },
-      "epoch": "2017-03-21T13:00:00.000Z",
-      "fees": {
-        "dynamic": true,
-        "dynamicFees": {
-          "minFeePool": 1000,
-          "minFeeBroadcast": 1000,
-          "addonBytes": {
-            "transfer": 100,
-            "secondSignature": 250,
-            "delegateRegistration": 400000,
-            "vote": 100,
-            "multiSignature": 500,
-            "ipfs": 250,
-            "timelockTransfer": 500,
-            "multiPayment": 500,
-            "delegateResignation": 400000
-          }
+    "data": {
+        "core": {
+            "version": "2.6.0-next.4"
         },
-        "staticFees": {
-          "transfer": 10000000,
-          "secondSignature": 500000000,
-          "delegateRegistration": 2500000000,
-          "vote": 100000000,
-          "multiSignature": 500000000,
-          "ipfs": 0,
-          "timelockTransfer": 0,
-          "multiPayment": 0,
-          "delegateResignation": 0
+        "nethash": "2a44f340d76ffc3df204c5f38cd355b7496c9065a1ade2ef92071436bd72e867",
+        "slip44": 1,
+        "wif": 170,
+        "token": "DARK",
+        "symbol": "DѦ",
+        "explorer": "https://dexplorer.ark.io",
+        "version": 30,
+        "ports": {
+            "@arkecosystem/core-p2p": 4002,
+            "@arkecosystem/core-api": 4003
+        },
+        "constants": {
+            "height": 4006000,
+            "reward": 200000000,
+            "activeDelegates": 51,
+            "blocktime": 8,
+            "block": {
+                "version": 0,
+                "maxTransactions": 500,
+                "maxPayload": 21000000,
+                "acceptExpiredTransactionTimestamps": false,
+                "idFullSha256": true
+            },
+            "epoch": "2017-03-21T13:00:00.000Z",
+            "fees": {
+                "staticFees": {
+                    "transfer": 10000000,
+                    "secondSignature": 500000000,
+                    "delegateRegistration": 2500000000,
+                    "vote": 100000000,
+                    "multiSignature": 500000000,
+                    "ipfs": 500000000,
+                    "multiPayment": 10000000,
+                    "delegateResignation": 2500000000,
+                    "htlcLock": 10000000,
+                    "htlcClaim": 0,
+                    "htlcRefund": 0
+                }
+            },
+            "ignoreInvalidSecondSignatureField": false,
+            "ignoreExpiredTransactions": false,
+            "vendorFieldLength": 255,
+            "multiPaymentLimit": 128,
+            "p2p": {
+                "minimumVersions": [
+                    "^2.6 || ^2.6.0-next.4"
+                ]
+            },
+            "aip11": true
+        },
+        "transactionPool": {
+            "dynamicFees": {
+                "enabled": true,
+                "minFeePool": 1000,
+                "minFeeBroadcast": 1000,
+                "addonBytes": {
+                    "transfer": 100,
+                    "secondSignature": 250,
+                    "delegateRegistration": 400000,
+                    "vote": 100,
+                    "multiSignature": 500,
+                    "ipfs": 250,
+                    "multiPayment": 500,
+                    "delegateResignation": 100,
+                    "htlcLock": 100,
+                    "htlcClaim": 0,
+                    "htlcRefund": 0
+                }
+            }
         }
-      }
-    },
-    "feeStatistics": [{
-      "type": 2,
-      "fees": {
-        "minFee": 100000,
-        "maxFee": 2500000000,
-        "avgFee": 2248696335
-      }
-    }, {
-      "type": 0,
-      "fees": {
-        "minFee": 1,
-        "maxFee": 1000000000000,
-        "avgFee": 11208027
-      }
-    }, {
-      "type": 1,
-      "fees": {
-        "minFee": 500000000,
-        "maxFee": 500000000,
-        "avgFee": 500000000
-      }
-    }, {
-      "type": 3,
-      "fees": {
-        "minFee": 1000000,
-        "maxFee": 100000000,
-        "avgFee": 99462267
-      }
-    }, {
-      "type": 4,
-      "fees": {
-        "minFee": 2000000000,
-        "maxFee": 2000000000,
-        "avgFee": 2000000000
-      }
-    }]
-  }
+    }
 }
