+++ tests/fixtures/node-status.json
@@ -1,7 +1,8 @@
 {
-  "data": {
-    "synced": true,
-    "now": 3940,
-    "blocksCount": -1
-  }
+    "data": {
+        "synced": true,
+        "now": 4168886,
+        "blocksCount": 0,
+        "timestamp": 83720178
+    }
 }
