+++ tests/fixtures/node-syncing.json
@@ -1,8 +1,8 @@
 {
-  "data": {
-    "syncing": false,
-    "blocks": -2,
-    "height": 3940,
-    "id": "17142334154459441029"
-  }
+    "data": {
+        "syncing": false,
+        "blocks": -2,
+        "height": 4168891,
+        "id": "74ce14f5994c44dfa3349568525e6c0a0d94b852913b1433e178611c8507089c"
+    }
 }
